pcli(dex_utils): forward sub tweak
